backend/requirements.txt
Django
djangorestframework
django-cors-headers
django-celery-beat==2.1.0
django-celery-results
django-extensions
websockets
celery
redis
channels
channels-redis
django-redis
psycopg2
Faker
GitPython
radon
django-jazzmin

PyGithub
lizard

pytest
pytest-django
pytest-mock

factory_boy
